requirements/ground.txt
CherryPy==3.2.4
Django==1.6.2
Jinja2==2.7.2
Markdown==2.4
appdirs==1.2.0
django-bootstrap3==2.6.1
django-filter==0.7
django-registration==1.0
djangorestframework==2.3.12
gevent==1.0
greenlet==0.4.2
gunicorn==18.0
jsonfield==0.9.20
mock==1.0.1
-e git+https://github.com/marshall/pynmea@39bfc18ede6d0ed898e329f844e7d0b28926482c#egg=pynmea-master
pyserial==2.7
